+++ relay-quotas/src/redis.rs
@@ -346,12 +346,24 @@ impl<T: GlobalLimiter> RedisRateLimiter<T> {
         // client across await points, otherwise it might be held for too long, and we will run out
         // of connections.
         let mut connection = self.client.get_connection().await?;
-        let rejections: Vec<bool> = invocation
+        let results: Vec<i64> = invocation
             .invoke_async(&mut connection)
             .await
             .map_err(RedisError::Redis)?;
 
-        for (quota, is_rejected) in tracked_quotas.iter().zip(rejections) {
+        let quantity = i64::try_from(quantity).unwrap_or(i64::MAX);
+        for (quota, consumed) in tracked_quotas.iter().zip(results) {
+            let limit = quota.limit();
+
+            // Limit must not be unlimited (-1) and there must not be enough remaining in the quota.
+            //
+            // Note: this logic is synchronized to the one in the Redis script.
+            let is_rejected = limit >= 0
+                && match quantity == 0 || over_accept_once {
+                    true => consumed >= limit,
+                    false => consumed + quantity > limit,
+                };
+
             if is_rejected {
                 let retry_after = self.retry_after((quota.expiry() - timestamp).as_secs());
                 rate_limits.add(RateLimit::from_quota(quota, *item_scoping, retry_after));
@@ -1003,7 +1015,6 @@ mod tests {
     }
 
     #[tokio::test]
-    #[allow(clippy::disallowed_names, clippy::let_unit_value)]
     async fn test_is_rate_limited_script() {
         let now = SystemTime::now()
             .duration_since(UNIX_EPOCH)
@@ -1039,22 +1050,35 @@ mod tests {
             .arg(1) // quantity
             .arg(false); // over accept once
 
-        // The item should not be rate limited by either key.
+        // Craft a new invocation similar to the previous one, but it only applies to the quota
+        // with a higher limit (2).
+        let mut invocation2 = script.prepare_invoke();
+        invocation2
+            .key(&bar) // key
+            .key(&r_bar) // refund key
+            .arg(2) // limit
+            .arg(now + 120) // expiry
+            .arg(1) // quantity
+            .arg(false); // over accept once
+
+        // Current usage is 0. But current values are now incremented by 1 (quantity).
         assert_eq!(
             invocation
-                .invoke_async::<Vec<bool>>(&mut conn)
+                .invoke_async::<Vec<i64>>(&mut conn)
                 .await
                 .unwrap(),
-            vec![false, false]
+            vec![0, 0]
         );
 
-        // The item should be rate limited by the first key (1).
+        // The usage was incremented in the last invocation, this invocation fails the rate limit
+        // on the first quota. -> No changes are made to the counters, the next invocation still
+        // needs to be `[1, 1]`.
         assert_eq!(
             invocation
-                .invoke_async::<Vec<bool>>(&mut conn)
+                .invoke_async::<Vec<i64>>(&mut conn)
                 .await
                 .unwrap(),
-            vec![true, false]
+            vec![1, 1]
         );
 
         // The item should still be rate limited by the first key (1), but *not*
@@ -1063,10 +1087,39 @@ mod tests {
         // quota don't affect unrelated items that share a parent quota.
         assert_eq!(
             invocation
-                .invoke_async::<Vec<bool>>(&mut conn)
+                .invoke_async::<Vec<i64>>(&mut conn)
                 .await
                 .unwrap(),
-            vec![true, false]
+            vec![1, 1]
+        );
+
+        // Using the second invocation which only considers a quota with a higher limit, this
+        // should still yield the current value of `1` and the next invocation should yield `2`.
+        assert_eq!(
+            invocation2
+                .invoke_async::<Vec<i64>>(&mut conn)
+                .await
+                .unwrap(),
+            vec![1]
+        );
+
+        // This now yields `2`. This is also the invocation at the limit, which means it should no
+        // longer increment the counter.
+        assert_eq!(
+            invocation2
+                .invoke_async::<Vec<i64>>(&mut conn)
+                .await
+                .unwrap(),
+            vec![2]
+        );
+
+        // Check again with the original invocation, this now yields `[1, 2]`.
+        assert_eq!(
+            invocation
+                .invoke_async::<Vec<i64>>(&mut conn)
+                .await
+                .unwrap(),
+            vec![1, 2]
         );
 
         assert_eq!(conn.get::<_, String>(&foo).await.unwrap(), "1");
@@ -1074,7 +1127,7 @@ mod tests {
         assert!(ttl >= 59);
         assert!(ttl <= 60);
 
-        assert_eq!(conn.get::<_, String>(&bar).await.unwrap(), "1");
+        assert_eq!(conn.get::<_, String>(&bar).await.unwrap(), "2");
         let ttl: u64 = conn.ttl(&bar).await.unwrap();
         assert!(ttl >= 119);
         assert!(ttl <= 120);
@@ -1095,22 +1148,31 @@ mod tests {
             .arg(1) // quantity
             .arg(false);
 
-        // increment
+        // increment, current quota is 0.
         assert_eq!(
             invocation
-                .invoke_async::<Vec<bool>>(&mut conn)
+                .invoke_async::<Vec<i64>>(&mut conn)
                 .await
                 .unwrap(),
-            vec![false]
+            vec![0]
         );
 
-        // test that it's rate limited without refund
+        // test that it's rate limited without refund.
         assert_eq!(
             invocation
-                .invoke_async::<Vec<bool>>(&mut conn)
+                .invoke_async::<Vec<i64>>(&mut conn)
                 .await
                 .unwrap(),
-            vec![true]
+            vec![1]
+        );
+
+        // Make sure, the counter wasn't incremented.
+        assert_eq!(
+            invocation
+                .invoke_async::<Vec<i64>>(&mut conn)
+                .await
+                .unwrap(),
+            vec![1]
         );
 
         let mut invocation = script.prepare_invoke();
@@ -1125,10 +1187,10 @@ mod tests {
         // test that refund key is used
         assert_eq!(
             invocation
-                .invoke_async::<Vec<bool>>(&mut conn)
+                .invoke_async::<Vec<i64>>(&mut conn)
                 .await
                 .unwrap(),
-            vec![false]
+            vec![-4]
         );
     }
 }